pyproject.toml
[project]
name = "csv_to_ical"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
    "icalendar==6.3.2",
    "uuid==1.30",
    "python-slugify==8.0.4"
]